refactor: move force-update-screen into screen-components
